pyproject.toml
[project]
name = "cryptic-crossword-eval"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "inspect-ai>=0.3.160",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
